added MainTabs to Navigation module
